rename TS kernel files and classes
