+++ yarn-project/circuits.js/src/structs/index.ts
@@ -12,13 +12,16 @@ export * from './global_variables.js';
 export * from './header.js';
 export * from './kernel/combined_accumulated_data.js';
 export * from './kernel/combined_constant_data.js';
-export * from './kernel/previous_kernel_data.js';
-export * from './kernel/private_kernel.js';
-export * from './kernel/public_inputs.js';
-export * from './kernel/public_inputs_final.js';
-export * from './kernel/public_kernel.js';
+export * from './kernel/private_kernel_inner_circuit_public_inputs.js';
+export * from './kernel/private_kernel_inner_data.js';
+export * from './kernel/private_kernel_tail_circuit_public_inputs.js';
+export * from './kernel/private_kernel_tail_data.js';
+export * from './kernel/public_kernel_circuit_private_inputs.js';
+export * from './kernel/public_kernel_circuit_public_inputs.js';
+export * from './kernel/public_kernel_data.js';
 export * from './membership_witness.js';
 export * from './nullifier_key_validation_request.js';
+export * from './partial_state_reference.js';
 export * from './private_call_stack_item.js';
 export * from './private_circuit_public_inputs.js';
 export * from './proof.js';
@@ -32,14 +35,13 @@ export * from './rollup/base_rollup.js';
 export * from './rollup/merge_rollup.js';
 export * from './rollup/previous_rollup_data.js';
 export * from './rollup/root_rollup.js';
+export * from './rollup/state_diff_hints.js';
 export * from './shared.js';
 export * from './side_effects.js';
-export * from './rollup/state_diff_hints.js';
+export * from './state_reference.js';
 export * from './tx_context.js';
 export * from './tx_request.js';
 export * from './verification_key.js';
-export * from './state_reference.js';
-export * from './partial_state_reference.js';
 
 export { FunctionSelector } from '@aztec/foundation/abi';
 export * from '@aztec/foundation/aztec-address';
