+++ yarn-project/circuits.js/src/structs/kernel/private_kernel_tail_data.ts
@@ -6,17 +6,17 @@ import { VK_TREE_HEIGHT } from '../../constants.gen.js';
 import { Proof, makeEmptyProof } from '../proof.js';
 import { UInt32 } from '../shared.js';
 import { VerificationKey } from '../verification_key.js';
-import { KernelCircuitPublicInputs } from './public_inputs.js';
+import { PrivateKernelTailCircuitPublicInputs } from './private_kernel_tail_circuit_public_inputs.js';
 
 /**
  * Data of the previous kernel iteration in the chain of kernels.
  */
-export class PreviousKernelData {
+export class PrivateKernelTailData {
   constructor(
     /**
      * Public inputs of the previous kernel.
      */
-    public publicInputs: KernelCircuitPublicInputs,
+    public publicInputs: PrivateKernelTailCircuitPublicInputs,
     /**
      * Proof of the previous kernel.
      */
@@ -35,23 +35,10 @@ export class PreviousKernelData {
     public vkPath: Tuple<Fr, typeof VK_TREE_HEIGHT>,
   ) {}
 
-  /**
-   * Serialize this as a buffer.
-   * @returns The buffer.
-   */
-  toBuffer() {
-    return serializeToBuffer(this.publicInputs, this.proof, this.vk, this.vkIndex, this.vkPath);
-  }
-
-  /**
-   * Deserializes from a buffer or reader, corresponding to a write in cpp.
-   * @param buffer - Buffer or reader to read from.
-   * @returns The deserialized instance.
-   */
-  static fromBuffer(buffer: Buffer | BufferReader): PreviousKernelData {
+  static fromBuffer(buffer: Buffer | BufferReader): PrivateKernelTailData {
     const reader = BufferReader.asReader(buffer);
     return new this(
-      reader.readObject(KernelCircuitPublicInputs),
+      reader.readObject(PrivateKernelTailCircuitPublicInputs),
       reader.readObject(Proof),
       reader.readObject(VerificationKey),
       reader.readNumber(),
@@ -59,17 +46,21 @@ export class PreviousKernelData {
     );
   }
 
-  /**
-   * Creates an empty instance, valid enough to be accepted by circuits.
-   * @returns The empty instance.
-   */
-  static empty(): PreviousKernelData {
-    return new PreviousKernelData(
-      KernelCircuitPublicInputs.empty(),
+  static empty(): PrivateKernelTailData {
+    return new this(
+      PrivateKernelTailCircuitPublicInputs.empty(),
       makeEmptyProof(),
       VerificationKey.makeFake(),
       0,
       makeTuple(VK_TREE_HEIGHT, Fr.zero),
     );
+  }
+
+  /**
+   * Serialize this as a buffer.
+   * @returns The buffer.
+   */
+  toBuffer() {
+    return serializeToBuffer(this.publicInputs, this.proof, this.vk, this.vkIndex, this.vkPath);
   }
 }
